Hit Animation
Units now flash white when hit. I had to implement this as a separate
textures because rendering is handled in the game class and it would be
too complicated to rework everything just for something small like this.
the sprites are pre loaded so it shouldn't be any more processor heavy
than the shader method Michel showed us.