Particles Update
Engine trail for the player has been implemented. I also fixed the
player bullet centering problem. It may happen on the enemies too but it
isn't noticeable. I also made it so the enemy's death particles appear
when the player collides with them. All these effects have also been
added to the first level as well. I also removed the level select render
message from the app log as it was cluttering the app log.